rag_poc_0203/pyproject.toml
[tool.poetry]
name = "ai-service-scaffold"
version = "0.1.0"
description = "Observability-First AI Service Architecture"
authors = ["AI Architect <architect@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
langchain = "^1.2.7"
langgraph = "^1.0.7"
langfuse = "^2.0.0"
openai = "^1.10.0"
qdrant-client = "^1.7.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.0"
langchain-openai = "^1.1.7"
streamlit = "^1.53.1"
langchain-ollama = "^1.0.1"
psycopg2-binary = "^2.9.11"
pdf4llm = "^0.2.9"
pdfplumber = "^0.11.9"
ragas = "^0.2.0"
datasets = "^3.2.0"
sqlalchemy = "^2.0.46"
aiosqlite = "^0.22.1"
greenlet = "^3.3.1"
ddgs = "^9.10.0"
lightrag-hku = "^1.4.9.11"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
